Changed the LONGs to float. The define in mathieeesingbas_intern.h does the trick.
git-svn-id: https://svn.aros.org/svn/aros/trunk@3961 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/mathieeesingbas/ieeespabs.c
@@ -17,10 +17,10 @@
 
     NAME */
 
-        AROS_LH1(LONG, IEEESPAbs,
+        AROS_LH1(float, IEEESPAbs,
 
 /*  SYNOPSIS */
-        AROS_LHA(LONG, y, D0),
+        AROS_LHA(float, y, D0),
 
 /*  LOCATION */
         struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 9, Mathieeesingbas)
