+++ rom/mathieeesingbas/ieeespceil.c
@@ -17,10 +17,10 @@
 
     NAME */
 
-        AROS_LH1(LONG, IEEESPCeil,
+        AROS_LH1(float, IEEESPCeil,
 
 /*  SYNOPSIS */
-        AROS_LHA(LONG, y, D0),
+        AROS_LHA(float, y, D0),
 
 /*  LOCATION */
         struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 16, Mathieeesingbas)
